real_estate/requirements.txt
Django
python-dotenv
python-telegram-bot
psycopg2-binary